Move transport traits to own module
