+++ proc-macros/src/lib.rs
@@ -327,7 +327,7 @@ fn build_api(api: api_def::ApiDefinition) -> Result<proc_macro2::TokenStream, sy
         quote_spanned!(api.name.span()=>
             #visibility async fn next_request(server: &'a mut jsonrpsee::core::RawServer<R, I>) -> core::result::Result<#enum_name #ty_generics, std::io::Error>
                 where
-                    R: jsonrpsee::core::TransportServer<RequestId = I>,
+                    R: jsonrpsee::transport::TransportServer<RequestId = I>,
                     I: Clone + PartialEq + Eq + std::hash::Hash + Send + Sync
                     #(, #params_tys: jsonrpsee::common::DeserializeOwned)*
             {
@@ -484,8 +484,8 @@ fn build_client_functions(
 
         client_functions.push(quote_spanned!(function.signature.span()=>
             // TODO: what if there's a conflict between `client` and a param name?
-            #visibility async fn #f_name<C: jsonrpsee::core::TransportClient>(client: &mut jsonrpsee::core::RawClient<C> #(, #params_list)*)
-                -> core::result::Result<#ret_ty, jsonrpsee::core::client::RawClientError<<C as jsonrpsee::core::TransportClient>::Error>>
+            #visibility async fn #f_name<C: jsonrpsee::transport::TransportClient>(client: &mut jsonrpsee::core::RawClient<C> #(, #params_list)*)
+                -> core::result::Result<#ret_ty, jsonrpsee::core::client::RawClientError<<C as jsonrpsee::transport::TransportClient>::Error>>
             where
                 #ret_ty: jsonrpsee::common::DeserializeOwned
                 #(, #params_tys: jsonrpsee::common::Serialize)*
