+++ src/core/server.rs
@@ -55,8 +55,6 @@ pub use self::notification::Notification;
 pub use self::params::{Iter as ParamsIter, ParamKey as ParamsKey, Params};
 pub use self::typed_rp::TypedResponder;
 
-pub mod raw;
-
 mod batch;
 mod batches;
 mod core;
