+++ src/server.rs
@@ -26,10 +26,10 @@
 
 use crate::common::{self, JsonValue};
 use crate::core::{
-    server::raw::TransportServer,
     server::{RawServerEvent, RawServerRequestId, RawServerSubscriptionId},
     RawServer,
 };
+use crate::transport::server::TransportServer;
 
 use futures::{channel::mpsc, future::Either, pin_mut, prelude::*};
 use parking_lot::Mutex;
