+++ src/transport/server.rs
@@ -33,8 +33,8 @@
 //! ## Example usage
 //!
 //! ```
-//! use jsonrpsee::core::server::raw::{TransportServer, TransportServerEvent};
 //! use jsonrpsee::common::{Error, Request, Response, Version};
+//! use jsonrpsee::transport::server::{TransportServer, TransportServerEvent};
 //!
 //! async fn run_server(server: &mut impl TransportServer) {
 //!     // Note that this implementation is a bit naive, as no request will be accepted by the
